Integrate boltdb for state and config data.
This is a big ol' commit, mostly because it's hard to isolate these
changes from each other.

Changes include but are not limited to:

* Command routing has been updated to accept a *bolt.DB value, giving
all commands access to state. I considered making it a *Bucket, but at
the moment is feels more logical to let everyone just grab their own
buckets. If an individual command needs both the `conf` bucket AND a
bucket for guild/channel state, *Bucket would be restrictive.

* updateState now populates buckets/keys instead of an in-memory map.

* LoadConfig no longer returns a Conf value, it just takes the *DB and
updates accordingly. The DB is now the source of truth for the
application, and Conf is only used to deserialize the config file.

* Added a handler to GuildCreate to create state buckets if necessary.